+++ lib/script-event.ts
@@ -1,51 +1,49 @@
 import { Block } from "./block.js";
 import { Entity } from "./entity.js";
 import { Player } from "./player.js";
-import { MessageSourceType } from "@minecraft/server";
+import { Wrapper } from "./wrapper.js";
+import { ScriptEventSource } from "@minecraft/server";
 import * as MC from "@minecraft/server";
 
-export { MessageSourceType };
+export { ScriptEventSource };
 
-export class ScriptEventCommandMessageEvent {
-    readonly #event: MC.ScriptEventCommandMessageEvent;
-
-    /** The constructor is public only because of a language
-     * limitation. User code must never call it directly. */
-    public constructor(rawEvent: MC.ScriptEventCommandMessageEvent) {
-        this.#event = rawEvent;
+export class ScriptEventCommandMessageAfterEvent extends Wrapper<MC.ScriptEventCommandMessageAfterEvent> {
+    /** Package private: user code should not use this. */
+    public constructor(rawEvent: MC.ScriptEventCommandMessageAfterEvent) {
+        super(rawEvent);
     }
 
     public get id(): string {
-        return this.#event.id;
+        return this.raw.id;
     }
 
     public get initiator(): Entity|undefined {
-        return this.#event.initiator
-            ? (this.#event.initiator.typeId === "minecraft:player"
-                ? new Player(this.#event.initiator as MC.Player)
-                : new Entity(this.#event.initiator))
+        return this.raw.initiator
+            ? (this.raw.initiator.typeId === "minecraft:player"
+                ? new Player(this.raw.initiator as MC.Player)
+                : new Entity(this.raw.initiator))
             : undefined;
     }
 
     public get message(): string {
-        return this.#event.message;
+        return this.raw.message;
     }
 
     public get sourceBlock(): Block|undefined {
-        return this.#event.sourceBlock
-            ? new Block(this.#event.sourceBlock)
+        return this.raw.sourceBlock
+            ? new Block(this.raw.sourceBlock)
             : undefined;
     }
 
     public get sourceEntity(): Entity|undefined {
-        return this.#event.sourceEntity
-            ? (this.#event.sourceEntity.typeId === "minecraft:player"
-                ? new Player(this.#event.sourceEntity as MC.Player)
-                : new Entity(this.#event.sourceEntity))
+        return this.raw.sourceEntity
+            ? (this.raw.sourceEntity.typeId === "minecraft:player"
+                ? new Player(this.raw.sourceEntity as MC.Player)
+                : new Entity(this.raw.sourceEntity))
             : undefined;
     }
 
-    public get sourceType(): MessageSourceType {
-        return this.#event.sourceType;
+    public get sourceType(): ScriptEventSource {
+        return this.raw.sourceType;
     }
 }