+++ lib/system.ts
@@ -1,15 +1,19 @@
-import { SystemEvents } from "./system/events.js";
+import { SystemAfterEvents, SystemBeforeEvents } from "./system/events.js";
 import { Wrapper } from "./wrapper.js";
 import * as MC from "@minecraft/server";
 
 export class System extends Wrapper<MC.System> {
-    /** System event signals */
-    public readonly events: SystemEvents;
+    /** System after-event signals */
+    public readonly afterEvents: SystemAfterEvents;
+
+    /** System before-event signals */
+    public readonly beforeEvents: SystemBeforeEvents;
 
     /** Package private: user code should not use this. */
     public constructor(rawSystem: MC.System) {
         super(rawSystem);
-        this.events = new SystemEvents(this.raw.events);
+        this.afterEvents  = new SystemAfterEvents(this.raw.afterEvents);
+        this.beforeEvents = new SystemBeforeEvents(this.raw.beforeEvents);
     }
 }