+++ lib/system/events.ts
@@ -1,27 +1,36 @@
 import { IEventSignal, GluedEventSignalWithoutOptions,
          GluedEventSignalWithOptions } from "../event.js";
-import { ScriptEventCommandMessageEvent } from "../script-event.js";
-import { BeforeWatchdogTerminateEvent } from "../watchdog.js";
+import { ScriptEventCommandMessageAfterEvent } from "../script-event.js";
+import { WatchdogTerminateBeforeEvent } from "../watchdog.js";
 import { Wrapper } from "../wrapper.js";
 import { identity } from "../function.js";
 import * as MC from "@minecraft/server";
 
-export class SystemEvents extends Wrapper<MC.SystemEvents> {
-    public readonly beforeWatchdogTerminate: IEventSignal<BeforeWatchdogTerminateEvent>;
-    public readonly scriptEventReceive:      IEventSignal<ScriptEventCommandMessageEvent, MC.ScriptEventMessageFilterOptions>;
+export class SystemAfterEvents extends Wrapper<MC.SystemAfterEvents> {
+    public readonly scriptEventReceive: IEventSignal<ScriptEventCommandMessageAfterEvent, MC.ScriptEventMessageFilterOptions>;
 
-    public constructor(rawEvents: MC.SystemEvents) {
+    /** Package private: user code should not use this. */
+    public constructor(rawEvents: MC.SystemAfterEvents) {
         super(rawEvents);
-        this.beforeWatchdogTerminate = new GluedEventSignalWithoutOptions(
-            this.raw.beforeWatchdogTerminate,
-            (rawEv: MC.BeforeWatchdogTerminateEvent) => {
-                return new BeforeWatchdogTerminateEvent(rawEv);
-            });
         this.scriptEventReceive = new GluedEventSignalWithOptions(
             this.raw.scriptEventReceive,
-            (rawEv: MC.ScriptEventCommandMessageEvent) => {
-                return new ScriptEventCommandMessageEvent(rawEv);
+            (rawEv: MC.ScriptEventCommandMessageAfterEvent) => {
+                return new ScriptEventCommandMessageAfterEvent(rawEv);
             },
             identity);
+    }
+}
+
+export class SystemBeforeEvents extends Wrapper<MC.SystemBeforeEvents> {
+    public readonly watchdogTerminate: IEventSignal<WatchdogTerminateBeforeEvent>;
+
+    /** Package private: user code should not use this. */
+    public constructor(rawEvents: MC.SystemBeforeEvents) {
+        super(rawEvents);
+        this.watchdogTerminate = new GluedEventSignalWithoutOptions(
+            this.raw.watchdogTerminate,
+            (rawEv: MC.WatchdogTerminateBeforeEvent) => {
+                return new WatchdogTerminateBeforeEvent(rawEv);
+            });
     }
 }