+++ lib/ui.ts
@@ -1,4 +1,7 @@
 import { Player } from "./player.js";
+import { Wrapper } from "./wrapper.js";
+import { RawMessage } from "@minecraft/server";
+import * as MC from "@minecraft/server";
 import * as UI from "@minecraft/server-ui";
 
 export {
@@ -9,109 +12,103 @@ export {
     ModalFormResponse
 } from "@minecraft/server-ui";
 
-export class ActionFormData {
-    readonly #raw: UI.ActionFormData;
-
+export class ActionFormData extends Wrapper<UI.ActionFormData> {
     public constructor() {
-        this.#raw = new UI.ActionFormData();
+        super(new UI.ActionFormData());
     }
 
-    body(bodyText: string): ActionFormData {
-        this.#raw.body(bodyText);
+    body(bodyText: RawMessage | string): ActionFormData {
+        this.raw.body(bodyText);
         return this;
     }
 
-    button(text: string, iconPath?: string): ActionFormData {
-        this.#raw.button(text, iconPath);
+    button(text: MC.RawMessage, iconPath?: string): ActionFormData {
+        this.raw.button(text, iconPath);
         return this;
     }
 
     show(player: Player): Promise<UI.ActionFormResponse> {
-        return this.#raw.show(player.rawPlayer);
+        return this.raw.show(player.rawPlayer);
     }
 
-    title(titleText: string): ActionFormData {
-        this.#raw.title(titleText);
+    title(titleText: MC.RawMessage | string): ActionFormData {
+        this.raw.title(titleText);
         return this;
     }
 }
 
-export class MessageFormData {
-    readonly #raw: UI.MessageFormData;
-
+export class MessageFormData extends Wrapper<UI.MessageFormData> {
     public constructor() {
-        this.#raw = new UI.MessageFormData();
+        super(new UI.MessageFormData());
     }
 
-    body(bodyText: string): MessageFormData {
-        this.#raw.body(bodyText);
+    body(bodyText: MC.RawMessage | string): MessageFormData {
+        this.raw.body(bodyText);
         return this;
     }
 
-    button1(text: string): MessageFormData {
-        this.#raw.button1(text);
+    button1(text: MC.RawMessage | string): MessageFormData {
+        this.raw.button1(text);
         return this;
     }
 
-    button2(text: string): MessageFormData {
-        this.#raw.button2(text);
+    button2(text: MC.RawMessage | string): MessageFormData {
+        this.raw.button2(text);
         return this;
     }
 
     show(player: Player): Promise<UI.MessageFormResponse> {
-        return this.#raw.show(player.rawPlayer);
+        return this.raw.show(player.rawPlayer);
     }
 
-    title(titleText: string): MessageFormData {
-        this.#raw.title(titleText);
+    title(titleText: MC.RawMessage | string): MessageFormData {
+        this.raw.title(titleText);
         return this;
     }
 }
 
-export class ModalFormData {
-    readonly #raw: UI.ModalFormData;
-
+export class ModalFormData extends Wrapper<UI.ModalFormData> {
     public constructor() {
-        this.#raw = new UI.ModalFormData();
+        super(new UI.ModalFormData());
     }
 
-    dropdown(label: string, options: string[], defaultValueIndex?: number): ModalFormData {
-        this.#raw.dropdown(label, options, defaultValueIndex);
-        return this;
-    }
-
-    icon(iconPath: string): ModalFormData {
-        this.#raw.icon(iconPath);
+    dropdown(label: MC.RawMessage | string,
+             options: (MC.RawMessage | string)[],
+             defaultValueIndex?: number
+            ): ModalFormData {
+        this.raw.dropdown(label, options, defaultValueIndex);
         return this;
     }
 
     show(player: Player): Promise<UI.ModalFormResponse> {
-        return this.#raw.show(player.rawPlayer);
+        return this.raw.show(player.rawPlayer);
     }
 
-    slider(
-        label: string,
-        minimumValue: number,
-        maximumValue: number,
-        valueStep: number,
-        defaultValue?: number,
-    ): ModalFormData {
-        this.#raw.slider(label, minimumValue, maximumValue, valueStep, defaultValue);
+    slider(label: MC.RawMessage | string,
+           minimumValue: number,
+           maximumValue: number,
+           valueStep: number,
+           defaultValue?: number,
+          ): ModalFormData {
+        this.raw.slider(label, minimumValue, maximumValue, valueStep, defaultValue);
         return this;
     }
 
-    textField(label: string, placeholderText: string, defaultValue?: string): ModalFormData {
-        this.#raw.textField(label, placeholderText, defaultValue);
+    textField(label: MC.RawMessage | string,
+              placeholderText: MC.RawMessage | string,
+              defaultValue?: string
+             ): ModalFormData {
+        this.raw.textField(label, placeholderText, defaultValue);
         return this;
     }
 
-    title(titleText: string): ModalFormData {
-        this.#raw.title(titleText);
+    title(titleText: MC.RawMessage | string): ModalFormData {
+        this.raw.title(titleText);
         return this;
     }
 
-    toggle(label: string, defaultValue?: boolean): ModalFormData {
-        this.#raw.toggle(label, defaultValue);
+    toggle(label: MC.RawMessage | string, defaultValue?: boolean): ModalFormData {
+        this.raw.toggle(label, defaultValue);
         return this;
     }
 }