+++ lib/world.ts
@@ -1,6 +1,6 @@
-import { WorldEvents } from "./world/events.js";
+import { WorldAfterEvents, WorldBeforeEvents } from "./world/events.js";
 import { map } from "./iterable.js";
-import { Player, PlayerSpawnEvent } from "./player.js"
+import { Player, PlayerSpawnAfterEvent } from "./player.js"
 import { HasDynamicProperties } from "./dynamic-props.js";
 import { Wrapper } from "./wrapper.js";
 import { MessageType } from "@protobuf-ts/runtime";
@@ -16,8 +16,11 @@ export class World extends HasDynamicProperties(Wrapper<MC.World>) implements IP
     #readinessProbe: number|null;
     #pendingEvents: (() => void)[];
 
-    /** World event signals */
-    public readonly events: WorldEvents;
+    /** World after-event signals */
+    public readonly afterEvents: WorldAfterEvents;
+
+    /** World before-event signals */
+    public readonly beforeEvents: WorldBeforeEvents;
 
     /** The constructor is public only because of a language
      * limitation. User code must never call it directly. */
@@ -26,7 +29,8 @@ export class World extends HasDynamicProperties(Wrapper<MC.World>) implements IP
         this.#isReady        = false;
         this.#readinessProbe = null;
         this.#pendingEvents  = [];
-        this.events          = new WorldEvents(this.raw.events);
+        this.afterEvents     = new WorldAfterEvents(this.raw.afterEvents);
+        this.beforeEvents    = new WorldBeforeEvents(this.raw.beforeEvents);
         this.#glueEvents();
     }
 
@@ -67,7 +71,7 @@ export class World extends HasDynamicProperties(Wrapper<MC.World>) implements IP
                 // ready yet, which isn't even documented!!
                 if (it) {
                     this.#isReady = true;
-                    this.events.ready.signal({});
+                    this.afterEvents.ready.signal({});
 
                     for (const ev of this.#pendingEvents) {
                         ev();
@@ -81,28 +85,28 @@ export class World extends HasDynamicProperties(Wrapper<MC.World>) implements IP
         };
         this.#readinessProbe = MC.system.runInterval(onTick, 1);
 
-        this.raw.events.playerSpawn.subscribe(rawEv => {
-            const ev: PlayerSpawnEvent = {
+        this.raw.afterEvents.playerSpawn.subscribe(rawEv => {
+            const ev: PlayerSpawnAfterEvent = {
                 initialSpawn: rawEv.initialSpawn,
                 player:       new Player(rawEv.player)
             };
             if (this.#isReady) {
-                this.events.playerSpawn.signal(ev);
+                this.afterEvents.playerSpawn.signal(ev);
             }
             else {
                 this.#pendingEvents.push(() => {
-                    this.events.playerSpawn.signal(ev);
+                    this.afterEvents.playerSpawn.signal(ev);
                 });
             }
         });
 
-        this.raw.events.playerLeave.subscribe(ev => {
+        this.raw.afterEvents.playerLeave.subscribe(ev => {
             if (this.#isReady) {
-                this.events.playerLeave.signal(ev);
+                this.afterEvents.playerLeave.signal(ev);
             }
             else {
                 this.#pendingEvents.push(() => {
-                    this.events.playerLeave.signal(ev);
+                    this.afterEvents.playerLeave.signal(ev);
                 });
             }
         });
@@ -114,6 +118,6 @@ export const world = new World(MC.world);
 // The reason why this is here is to avoid circular imports. It's usually
 // not a problem but here we rely on code on the top level so we must avoid
 // that at all cost.
-world.events.worldInitialize.subscribe(ev => {
+world.afterEvents.worldInitialize.subscribe(ev => {
     Prefs.initialise(ev.propertyRegistry);
 });