+++ NEWS.md
@@ -2,12 +2,15 @@
 
 ## 2.0.0 -- not released yet
 
+* Updated for Minecraft Bedrock 1.20.0 (@minecraft/server 1.3.0-beta).
 * New properties:
   * `Block.prototype.offset`
   * `Entity.prototype.getBlockFromViewDirection`
   * `Player.prototype.console`
-  * `World.prototype.events`
-  * `System.prototype.events`
+  * `World.prototype.afterEvents`
+  * `World.prototype.beforeEvents`
+  * `System.prototype.afterEvents`
+  * `System.prototype.beforeEvents`
 * Not only `Player`, now `World` also supports `getDynamicProperty`,
   `setDynamicProperty`, `removeDynamicProperty`, `getPreferences`, and
   `setPreferences`.